requirements.txt
pandas
numpy
sqlalchemy
streamlit
plotly
python-dateutil
google-generativeai
streamlit-option-menu
bcrypt